Refactored DesignPanel into a single component
